Rename scripts files to be more CLI-friendly
- Use naming conventions based on scripting language used:
  snake_case for bash and kebab-case for Powershell.

Not like it matters for Powershell since it just cycles through the
files if they start with the same names. But for Linux it's a massive
pain to autocomplete the filenames if they start with a capital letter,
and have spaces in it. Either way, I should really name the files to be
more autocomplete-friendly.

- Convert compile_aomenc.sh's EOL sequence to LF

Took me a while to figure out why it's breaking in my WSL Ubuntu.
Although I think Git converted LF back to CRLF in the index, ugh.