Update `Compile av1an.md`
It's certainly nice to see more people attempting to compile Av1an on
Windows and creating guides for it, like https://redd.it/s8151l. It's
also great that the developers themselves plan to add a Windows
compilation guide (see issue #529 in master-of-zen/Av1an). In the
meantime, I'm just going to update the guide as if anyone will ever find
this guide by chance. :D