Update av1an compile script
The previous MSYS-based compile_av1an.sh script is now replaced with a
new Visual Studio-based Compile-Av1an.ps1 script, to reflect the
official compilation guide added in master-of-zen/Av1an@3675f05.

The biggest benefit to using the Visual Studio method is that I can now
compile av1an with ffmpeg 5.0. That's not possible with MSYS due to
msys2/MINGW-packages#10725.

I don't plan to remove Compile_av1an.md because it is still a viable
alternative. Both options are just as involved so pick your poison :)