Remove version checking for FFmpeg
BtbN no longer includes commit hashes in the filename for the latest
release; instead, it uses the "latest" tag. It makes things simpler I
suppose, but that means I won't be able to check the version before
downloading.

Since I only run the script weekly, and ffmpeg is bound to get updated
daily anyways, it's fine to just download the latest version every time
the script is run.